Modularise tech SQL queries into one file
