--- src/model/queries/getAllTechStack.js
@@ -1,7 +1,0 @@
-const dbConnection = require('../database/db_connection');
-
-const getAllTechStack = () => dbConnection.query(
-  'SELECT * FROM tech_stack'
-);
-
-module.exports = getAllTechStack;
